V1.02 Update: Bolt Flywheel Cage
Mounted with two M3x20 bolts in the center and one M3x30 bolt down the front (plus two heatset inserts and 1 hex nut), this is a flywheel cage option that should help dampen vibrations from your cage somewhat (a bit of an issue given the original pin mounting system seems prone to issues from vibration). It was mostly finished a while ago, but an error was made when modeling it and the cages needed to be re-exported.